requirements.txt
python-dotenv>=1.0.0
ddgs>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
